GC-Notify-Access Request functionality done, unit tests WIP
