+++ .env.local
@@ -40,6 +40,5 @@ GC_NOTIFY_CREATE_EMAIL_TEMPLATE_ID='create-email-template-id'
 GC_NOTIFY_CREATE_PHONE_TEMPLATE_ID='create-phone-template-id'
 GC_NOTIFY_ACCESS_REQUEST_EMAIL_TEMPLATE_ID='access-request-email-template-id'
 GC_NOTIFY_VHERS_ADMIN_EMAIL='vhers-admin-email'
-GC_NOTIFY_SUPPORT_STAFF_AND_ADMIN_EMAIL_LIST='support-staff-and-admin-email-list'
 
 VHERS_API_KEY=''
